CAF-5 Change dbcontext of user to roles
