+++ ControlAcceso.Tests/Data/UsersDbContextTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Data.Common;
 using ControlAcceso.Data.Model;
 using ControlAcceso.Data.Users;
+using ControlAcceso.Data.Roles;
 using ControlAcceso.Services.DBService;
 using Moq;
 
@@ -255,7 +256,7 @@ namespace ControlAcceso.Tests.Data
                 It.Is<Dictionary<string, object>>(parameters => (int)parameters["@RoleId"] == roleId)))
                 .Returns(expectedResult);  
 
-            var context = new UsersDbContext(_dbServiceMock.Object);
+            var context = new RolesDbContext(_dbServiceMock.Object);
 
             // Act
             var result = context.RoleExists(roleId);
