+++ ControlAcceso/Data/Roles/RolesDbContext.cs
@@ -112,6 +112,21 @@ public class RolesDbContext:IRolesDbContext
         }
     }
 
+    public bool RoleExists(int roleId)
+    {
+        var query = @"
+            SELECT 1 
+            FROM Roles 
+            WHERE id = @RoleId";
 
+        var parameters = new Dictionary<string, dynamic>
+        {
+            { "@RoleId", roleId }
+        };
 
+        var result = DbService.ExecuteReader(query, parameters);
+
+        return result != null && result.Any();
+
+    }
 }
