+++ ControlAcceso/Data/Users/IUsersDbContext.cs
@@ -20,7 +20,5 @@ namespace ControlAcceso.Data.Users
         
         public void UpdateUserRole(int idUser, int roleId);
 
-        public bool RoleExists(int roleId);
-
     }
 }
