+++ ControlAcceso/Data/Users/UsersDbContext.cs
@@ -193,23 +193,5 @@ namespace ControlAcceso.Data.Users
             }
         }
 
-        public bool RoleExists(int roleId)
-        {
-            var query = @"
-                SELECT 1 
-                FROM Roles 
-                WHERE id = @RoleId";
-
-            var parameters = new Dictionary<string, dynamic>
-            {
-                { "@RoleId", roleId }
-            };
-
-            var result = DbService.ExecuteReader(query, parameters);
-
-            return result != null && result.Any();
-
-        }
-
     }
 }
